Fix: 아키텍처 구조 수정 및 유저 정보 수정 api patch로 수정
