+++ src/main/java/elice/yeardreamback/service/UserService.java
@@ -9,29 +9,12 @@ import org.springframework.stereotype.Service;
 import java.util.Optional;
 
 @Service
-public class UserService {
+public interface UserService {
 
-    private final UserRepository userRepository;
-
-    public UserService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    public Optional<User> findUserByUsername(String username) {
-        return userRepository.findByUsername(username);
-    }
+    Optional<User> findUserByUsername(String username);
 
     @Transactional
-    public User updateUser(String username, String newName, String newRole, String newEmail, String newProfileImageUrl, String newPhone) {
-        User user = userRepository.findByUsername(username)
-                        .orElseThrow(() -> new UserNotFoundException(username));
+    User updateUser(String username, String newName, String newRole, String newEmail, String newProfileImageUrl, String newPhone);
 
-        user.setName(newName);
-        user.setRole(newRole);
-        user.setEmail(newEmail);
-        user.setProfileImg(newProfileImageUrl);
-        user.setPhone(newPhone);
-
-        return user;
-    }
+    Optional<User> logoutUser(String username);
 }
